Move util::logical_geometry to style
