+++ components/layout/flex.rs
@@ -27,10 +27,10 @@ use model::{IntrinsicISizes};
 use std::cmp::max;
 use std::sync::Arc;
 use style::computed_values::{flex_direction, float};
+use style::logical_geometry::LogicalSize;
 use style::properties::ComputedValues;
 use style::properties::style_structs;
 use style::values::computed::LengthOrPercentageOrAuto;
-use util::logical_geometry::LogicalSize;
 use util::opts;
 
 // A mode describes which logical axis a flex axis is parallel with.
