+++ components/layout/layout_thread.rs
@@ -65,6 +65,7 @@ use style::computed_values::{filter, mix_blend_mode};
 use style::context::{ReflowGoal, StylistWrapper};
 use style::dom::{TDocument, TElement, TNode};
 use style::error_reporting::ParseErrorReporter;
+use style::logical_geometry::LogicalPoint;
 use style::media_queries::{Device, MediaType};
 use style::parallel::WorkQueueData;
 use style::selector_impl::ServoSelectorImpl;
@@ -75,7 +76,6 @@ use traversal::RecalcStyleAndConstructFlows;
 use url::Url;
 use util::geometry::MAX_RECT;
 use util::ipc::OptionalIpcSender;
-use util::logical_geometry::LogicalPoint;
 use util::opts;
 use util::thread;
 use util::thread_state;
