+++ components/layout/table_colgroup.rs
@@ -15,9 +15,9 @@ use layout_debug;
 use std::cmp::max;
 use std::fmt;
 use std::sync::Arc;
+use style::logical_geometry::LogicalSize;
 use style::properties::ComputedValues;
 use style::values::computed::LengthOrPercentageOrAuto;
-use util::logical_geometry::LogicalSize;
 
 /// A table formatting context.
 pub struct TableColGroupFlow {
