+++ components/style/lib.rs
@@ -55,6 +55,7 @@ pub mod dom;
 pub mod element_state;
 pub mod error_reporting;
 pub mod font_face;
+pub mod logical_geometry;
 pub mod matching;
 pub mod media_queries;
 pub mod parallel;
